# apps/indigo-api-gateway/application/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API gateway for Indigo Server.",
        "title": "Indigo Server",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/organization/invites": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a paginated list of invites for the current organization.",
                "tags": [
                    "Administration API"
                ],
                "summary": "List Organization Invites",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cursor pointing to a record after which to fetch results",
                        "name": "after",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Maximum number of results to return",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved list of invites",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses_openai.ListResponse-app_interfaces_http_routes_v1_organization_invites_InviteResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid or missing query parameter",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing API key",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new invite for a user to join the organization.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Administration API"
                ],
                "summary": "Create Invite",
                "parameters": [
                    {
                        "description": "Invite request payload",
                        "name": "invite",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app_interfaces_http_routes_v1_organization_invites.CreateInviteUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully created invite",
                        "schema": {
                            "$ref": "#/definitions/app_interfaces_http_routes_v1_organization_invites.InviteResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload or user already exists",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing API key",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/organization/invites/verification": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Verifies an invitation code, checks expiration, registers the user if necessary, and assigns project memberships.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Administration API"
                ],
                "summary": "Verify Invite",
                "parameters": [
                    {
                        "description": "Verification request payload",
                        "name": "verification",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app_interfaces_http_routes_v1_organization_invites.VerifyInviteUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully verified invite",
                        "schema": {
                            "$ref": "#/definitions/app_interfaces_http_routes_v1_organization_invites.InviteResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid or expired invite code",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing API key",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/organization/invites/{invite_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a specific invite by its ID.",
                "tags": [
                    "Administration API"
                ],
                "summary": "Retrieve Invite",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Public ID of the invite",
                        "name": "invite_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved invite",
                        "schema": {
                            "$ref": "#/definitions/app_interfaces_http_routes_v1_organization_invites.InviteResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing API key",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Invite not found",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a specific invite by its ID. Only organization owners can delete invites.",
                "tags": [
                    "Administration API"
                ],
                "summary": "Delete Invite",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Public ID of the invite",
                        "name": "invite_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted invite",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses_openai.DeleteResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing API key",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - only owners can delete invites",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Invite not found",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "app_interfaces_http_routes_v1_organization_invites.CreateInviteUserRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "projects": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/app_interfaces_http_routes_v1_organization_invites.InviteProject"
                    }
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "app_interfaces_http_routes_v1_organization_invites.InviteProject": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "app_interfaces_http_routes_v1_organization_invites.InviteResponse": {
            "type": "object",
            "properties": {
                "accepted_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "invited_at": {
                    "type": "string"
                },
                "object": {
                    "type": "string"
                },
                "projects": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/app_interfaces_http_routes_v1_organization_invites.InviteProject"
                    }
                },
                "role": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "app_interfaces_http_routes_v1_organization_invites.VerifyInviteUserRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                }
            }
        },
...
        "menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "menlo_ai_indigo-api-gateway_app_interfaces_http_responses_openai.DeleteResponse": {
            "type": "object",
            "properties": {
                "deleted": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "object": {
                    "type": "string"
                }
            }
        },
...
        "menlo_ai_indigo-api-gateway_app_interfaces_http_responses_openai.ListResponse-app_interfaces_http_routes_v1_organization_invites_InviteResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/app_interfaces_http_routes_v1_organization_invites.InviteResponse"
                    }
                },
                "first_id": {
                    "type": "string"
                },
                "has_more": {
                    "type": "boolean"
                },
                "last_id": {
                    "type": "string"
                },
                "object": {
                    "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses_openai.ObjectTypeList"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "menlo_ai_indigo-api-gateway_app_interfaces_http_responses_openai.ObjectTypeList": {
            "type": "string",
            "enum": [
                "list"
            ],
            "x-enum-varnames": [
                "ObjectTypeListList"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}